@@ -1,24 +1,36 @@
 name: flutter_ume
-description: UME is an in-app debug kits platform for Flutter apps. Produced by Flutter Infra team of ByteDance.
-version: 1.1.2
-homepage: https://github.com/bytedance/flutter_ume
+description: UME is an in-app debug kits platform for Flutter apps. Produced by Flutter Infra team of hzcxfw.
+
+version: 0.1.11
+homepage: https://gitlab.hzcxfw.com/terminal/structure/flutter_ume
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
   flutter: ">=2.0.0"
+
 
 dependencies:
   flutter:
     sdk: flutter
   flutter_localizations:
     sdk: flutter
-
-  shared_preferences: ^2.0.15
-  vm_service: ">=9.4.0 <12.0.0"
+  touch_indicator: ^2.0.0
+  shared_preferences: ^2.2.2
+  share: ^2.0.4
+  vm_service: ^13.0.0
+  string_scanner: ^1.1.0
+  dio: ^5.4.0
+  system_info: ^1.0.1
+  device_info: ^2.0.2
+  platform: ^3.0.0
   tuple: ^2.0.0
+  rxdart: ^0.26.0
+  image: ^4.0.15
+  quiver: ^3.2.1
+  basics: ">=0.6.0 <0.11.0"
 
 dev_dependencies:
-  flutter_lints: any
+  flutter_lints: ^2.0.1
   flutter_test:
     sdk: flutter
   melos: ^3.1.0
@@ -28,6 +40,12 @@ dev_dependencies:
       url: https://github.com/smileShirely/flutter_coverage_badge.git
       ref: 59b7580f406bb712e9d9049c8c99212946e34f65
   lints: ^2.0.0
+#  image: ^4.0.15
+#  quiver: ^3.2.1
+#  basics: ^0.10.0
+
+dependency_overrides:
+  vm_service: ^13.0.0
 
 flutter:
   uses-material-design: true
