@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_ume/flutter_ume.dart';
+
 import 'custom_log.dart';
 
 void main() {
@@ -58,24 +59,20 @@ class _MyHomePageState extends State<MyHomePage> {
                 ),
                 Text(
                   '$_counter',
-                  style: Theme.of(context).textTheme.headline4,
+                  style: Theme.of(context).textTheme.headlineMedium,
                 ),
               ],
             ),
           ),
           Align(
               alignment: Alignment.bottomRight,
-              child: Row(
-                  mainAxisAlignment: MainAxisAlignment.end,
-                  children: <Widget>[
-                    Text(
-                      _hasTapped
-                          ? 'Open \nCustomLog \nto view log'
-                          : 'Tap here 👉',
-                      style: Theme.of(context).textTheme.headline4,
-                    ),
-                    SizedBox(width: 80)
-                  ])),
+              child: Row(mainAxisAlignment: MainAxisAlignment.end, children: <Widget>[
+                Text(
+                  _hasTapped ? 'Open \nCustomLog \nto view log' : 'Tap here 👉',
+                  style: Theme.of(context).textTheme.headlineMedium,
+                ),
+                SizedBox(width: 80)
+              ])),
         ],
       ),
       floatingActionButton: FloatingActionButton(
