@@ -50,8 +50,8 @@ class _DragableGridViewState<T> extends State<DragableGridView> {
         physics: BouncingScrollPhysics(),
         padding: EdgeInsets.zero,
         itemCount: dataList!.length,
-        gridDelegate: SliverGridDelegateWithFixedCrossAxisCount(
-            crossAxisCount: 3, childAspectRatio: 0.85),
+        gridDelegate:
+            SliverGridDelegateWithFixedCrossAxisCount(crossAxisCount: 3, childAspectRatio: 0.85),
         itemBuilder: ((ctx, index) {
           return _buildDraggable(ctx, index);
         }));
@@ -63,7 +63,7 @@ class _DragableGridViewState<T> extends State<DragableGridView> {
         return LongPressDraggable(
           data: index,
           child: DragTarget<int>(
-            onAccept: (_) {},
+            onAcceptWithDetails: (_) {},
             builder: (context, data, rejects) {
               return willAcceptIndex >= 0 && willAcceptIndex == index
                   ? Container()
@@ -76,13 +76,14 @@ class _DragableGridViewState<T> extends State<DragableGridView> {
                 dataList = dataListBackup.sublist(0);
               });
             },
-            onWillAccept: (int? fromIndex) {
+            onWillAcceptWithDetails: (DragTargetDetails<int> details) {
+              int fromIndex = details.data;
               final accept = fromIndex != index;
               if (accept) {
                 willAcceptIndex = index;
                 showItemWhenCovered = true;
                 dataList = dataListBackup.sublist(0);
-                final fromData = dataList![fromIndex!];
+                final fromData = dataList![fromIndex];
                 setState(() {
                   dataList!.removeAt(fromIndex);
                   dataList!.insert(index, fromData);
@@ -130,9 +131,7 @@ class _DragableGridViewState<T> extends State<DragableGridView> {
           ),
           childWhenDragging: Container(
             child: SizedBox(
-              child: showItemWhenCovered
-                  ? widget.itemBuilder(context, dataList![index])
-                  : null,
+              child: showItemWhenCovered ? widget.itemBuilder(context, dataList![index]) : null,
             ),
           ),
         );
