@@ -1,18 +1,19 @@
-import 'package:flutter/material.dart'
-    hide FlutterLogo, FlutterLogoDecoration, FlutterLogoStyle;
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart' hide FlutterLogo, FlutterLogoDecoration, FlutterLogoStyle;
+import 'package:flutter_localizations/flutter_localizations.dart';
+import 'package:flutter_ume/core/pluggable.dart';
 import 'package:flutter_ume/core/pluggable_message_service.dart';
-import 'package:flutter_ume/core/ui/panel_action_define.dart';
 import 'package:flutter_ume/core/plugin_manager.dart';
 import 'package:flutter_ume/core/red_dot.dart';
 import 'package:flutter_ume/core/store_manager.dart';
+import 'package:flutter_ume/core/ui/panel_action_define.dart';
 import 'package:flutter_ume/core/ui/toolbar_widget.dart';
-import 'package:flutter_ume/core/pluggable.dart';
 import 'package:flutter_ume/util/binding_ambiguate.dart';
 import 'package:flutter_ume/util/constants.dart';
-import './menu_page.dart';
 import 'package:flutter_ume/util/flutter_logo.dart';
+
+import './menu_page.dart';
 import 'global.dart';
-import 'package:flutter_localizations/flutter_localizations.dart';
 
 const defaultLocalizationsDelegates = const [
   GlobalMaterialLocalizations.delegate,
@@ -42,8 +43,7 @@ class UMEWidget extends StatefulWidget {
   /// The method does not have side-effects whether the [UMEWidget]
   /// is not enabled or no plugin has been activated.
   static void closeActivatedPlugin() {
-    final _ContentPageState? state =
-        _umeWidgetState?._contentPageKey.currentState;
+    final _ContentPageState? state = _umeWidgetState?._contentPageKey.currentState;
     if (state?._currentSelected != null) {
       state?._closeActivatedPluggable();
     }
@@ -84,9 +84,8 @@ class _UMEWidgetState extends State<UMEWidget> {
     _replaceChild();
     _injectOverlay();
 
-    _onMetricsChanged =
-        bindingAmbiguate(WidgetsBinding.instance)!.window.onMetricsChanged;
-    bindingAmbiguate(WidgetsBinding.instance)!.window.onMetricsChanged = () {
+    _onMetricsChanged = bindingAmbiguate(PlatformDispatcher.instance)!.onMetricsChanged;
+    bindingAmbiguate(PlatformDispatcher.instance)!.onMetricsChanged = () {
       if (_onMetricsChanged != null) {
         _onMetricsChanged!();
         _replaceChild();
@@ -98,8 +97,8 @@ class _UMEWidgetState extends State<UMEWidget> {
   @override
   void dispose() {
     if (_onMetricsChanged != null) {
-      bindingAmbiguate(WidgetsBinding.instance)!.window.onMetricsChanged =
-          _onMetricsChanged;
+      // PlatformDispatcher.instance.implicitView!
+      bindingAmbiguate(PlatformDispatcher.instance)!.onMetricsChanged = _onMetricsChanged;
     }
     super.dispose();
     // Do the cleaning at last.
@@ -124,12 +123,11 @@ class _UMEWidgetState extends State<UMEWidget> {
   }
 
   void _replaceChild() {
-    final nestedWidgets =
-        PluginManager.instance.pluginsMap.values.where((value) {
+    final nestedWidgets = PluginManager.instance.pluginsMap.values.where((value) {
       return value != null && value is PluggableWithNestedWidget;
     }).toList();
-    Widget layoutChild = _buildLayout(
-        widget.child, widget.supportedLocales, widget.localizationsDelegates);
+    Widget layoutChild =
+        _buildLayout(widget.child, widget.supportedLocales, widget.localizationsDelegates);
     for (var item in nestedWidgets) {
       if (item!.name != PluginManager.instance.activatedPluggableName) {
         continue;
@@ -139,18 +137,17 @@ class _UMEWidgetState extends State<UMEWidget> {
         break;
       }
     }
-    _child =
-        Directionality(textDirection: TextDirection.ltr, child: layoutChild);
+    _child = Directionality(textDirection: TextDirection.ltr, child: layoutChild);
   }
 
-  Stack _buildLayout(Widget child, Iterable<Locale>? supportedLocales,
-      Iterable<LocalizationsDelegate> delegates) {
+  Stack _buildLayout(
+      Widget child, Iterable<Locale>? supportedLocales, Iterable<LocalizationsDelegate> delegates) {
     return Stack(
       children: <Widget>[
         RepaintBoundary(child: child, key: rootKey),
         MediaQuery(
-          data: MediaQueryData.fromWindow(
-              bindingAmbiguate(WidgetsBinding.instance)!.window),
+          data:
+              MediaQueryData.fromView(bindingAmbiguate(PlatformDispatcher.instance)!.implicitView!),
           child: Localizations(
             locale: supportedLocales?.first ?? Locale('en', 'US'),
             delegates: delegates.toList(),
@@ -269,8 +266,7 @@ class _ContentPageState extends State<_ContentPage> {
 
   void _updatePanelWidget() {
     setState(() {
-      _currentWidget =
-          _showedMenu ? (_minimalContent ? _toolbarWidget : _menuPage) : _empty;
+      _currentWidget = _showedMenu ? (_minimalContent ? _toolbarWidget : _menuPage) : _empty;
     });
   }
 
@@ -284,9 +280,7 @@ class _ContentPageState extends State<_ContentPage> {
   Widget _logoWidget() {
     if (_currentSelected != null) {
       return Container(
-          child: Image(image: _currentSelected!.iconImageProvider),
-          height: 30,
-          width: 30);
+          child: Image(image: _currentSelected!.iconImageProvider), height: 30, width: 30);
     }
     return FlutterLogo(size: 40, colors: _showedMenu ? Colors.red : null);
   }
@@ -324,8 +318,7 @@ class _ContentPageState extends State<_ContentPage> {
       if (pluginData is PluggableWithAnywhereDoor) {
         dynamic result;
         if (pluginData.routeNameAndArgs != null) {
-          result = await pluginData.navigator?.pushNamed(
-              pluginData.routeNameAndArgs!.item1,
+          result = await pluginData.navigator?.pushNamed(pluginData.routeNameAndArgs!.item1,
               arguments: pluginData.routeNameAndArgs!.item2);
         } else if (pluginData.route != null) {
           result = await pluginData.navigator?.push(pluginData.route!);
@@ -375,8 +368,7 @@ class _ContentPageState extends State<_ContentPage> {
     _context = context;
     if (_windowSize.isEmpty) {
       _dx = MediaQuery.of(context).size.width - dotSize.width - margin * 4;
-      _dy =
-          MediaQuery.of(context).size.height - dotSize.height - bottomDistance;
+      _dy = MediaQuery.of(context).size.height - dotSize.height - bottomDistance;
       _windowSize = MediaQuery.of(context).size;
     }
     return Container(
@@ -419,9 +411,7 @@ class _ContentPageState extends State<_ContentPage> {
                           right: 6,
                           top: 8,
                           child: RedDot(
-                            pluginDatas: PluginManager
-                                .instance.pluginsMap.values
-                                .toList(),
+                            pluginDatas: PluginManager.instance.pluginsMap.values.toList(),
                           ))
                     ],
                   ),
