@@ -4,11 +4,9 @@
 
 import 'dart:math' as math;
 import 'dart:ui' as ui show Gradient, TextBox, lerpDouble, Color;
-import 'dart:typed_data';
 
 import 'package:flutter/foundation.dart';
-import 'package:flutter/material.dart'
-    hide FlutterLogo, FlutterLogoDecoration, FlutterLogoStyle;
+import 'package:flutter/material.dart' hide FlutterLogo, FlutterLogoDecoration, FlutterLogoStyle;
 
 /// The Flutter logo, in widget form. This widget respects the [IconTheme].
 /// For guidelines on using the Flutter logo, visit https://flutter.dev/brand.
@@ -128,8 +126,8 @@ class FlutterLogoDecoration extends Decoration {
         // (see https://github.com/dart-lang/sdk/issues/26980 for details about that ignore statement)
         _opacity = 1.0;
 
-  const FlutterLogoDecoration._(this.lightColor, this.darkColor, this.textColor,
-      this.style, this.margin, this._position, this._opacity);
+  const FlutterLogoDecoration._(this.lightColor, this.darkColor, this.textColor, this.style,
+      this.margin, this._position, this._opacity);
 
   /// The lighter of the two colors used to paint the logo.
   ///
@@ -168,13 +166,11 @@ class FlutterLogoDecoration extends Decoration {
 
   // The following are set when lerping, to represent states that can't be
   // represented by the constructor.
-  final double
-      _position; // -1.0 for stacked, 1.0 for horizontal, 0.0 for no logo
+  final double _position; // -1.0 for stacked, 1.0 for horizontal, 0.0 for no logo
   final double _opacity; // 0.0 .. 1.0
 
   bool get _inTransition =>
-      _opacity != 1.0 ||
-      (_position != -1.0 && _position != 0.0 && _position != 1.0);
+      _opacity != 1.0 || (_position != -1.0 && _position != 0.0 && _position != 1.0);
 
   @override
   bool debugAssertIsValid() {
@@ -204,8 +200,7 @@ class FlutterLogoDecoration extends Decoration {
   /// See also:
   ///
   ///  * [Decoration.lerp], which interpolates between arbitrary decorations.
-  static FlutterLogoDecoration? lerp(
-      FlutterLogoDecoration? a, FlutterLogoDecoration? b, double t) {
+  static FlutterLogoDecoration? lerp(FlutterLogoDecoration? a, FlutterLogoDecoration? b, double t) {
     assert(a == null || a.debugAssertIsValid());
     assert(b == null || b.debugAssertIsValid());
     if (a == null && b == null) return null;
@@ -267,8 +262,7 @@ class FlutterLogoDecoration extends Decoration {
   @override
   // ignore: todo
   // TODO(ianh): better hit testing
-  bool hitTest(Size size, Offset position, {TextDirection? textDirection}) =>
-      true;
+  bool hitTest(Size size, Offset position, {TextDirection? textDirection}) => true;
 
   @override
   BoxPainter createBoxPainter([VoidCallback? onChanged]) {
@@ -304,8 +298,7 @@ class FlutterLogoDecoration extends Decoration {
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
-    properties
-        .add(DiagnosticsNode.message('$lightColor/$darkColor on $textColor'));
+    properties.add(DiagnosticsNode.message('$lightColor/$darkColor on $textColor'));
     properties.add(EnumProperty<FlutterLogoStyle>('style', style));
     if (_inTransition) {
       properties.add(DiagnosticsNode.message(
@@ -347,11 +340,9 @@ class _FlutterLogoPainter extends BoxPainter {
     );
     _textPainter.layout();
     final ui.TextBox textSize = _textPainter
-        .getBoxesForSelection(
-            const TextSelection(baseOffset: 0, extentOffset: kLabel.length))
+        .getBoxesForSelection(const TextSelection(baseOffset: 0, extentOffset: kLabel.length))
         .single;
-    _textBoundingRect = Rect.fromLTRB(
-        textSize.left, textSize.top, textSize.right, textSize.bottom);
+    _textBoundingRect = Rect.fromLTRB(textSize.left, textSize.top, textSize.right, textSize.bottom);
   }
 
   // This class contains a lot of magic numbers. They were derived from the
@@ -371,8 +362,7 @@ class _FlutterLogoPainter extends BoxPainter {
     canvas.translate((202.0 - 166.0) / 2.0, 0.0);
 
     // Set up the styles.
-    final Paint lightPaint = Paint()
-      ..color = _config.lightColor!.withOpacity(0.8);
+    final Paint lightPaint = Paint()..color = _config.lightColor!.withOpacity(0.8);
     final Paint mediumPaint = Paint()..color = _config.lightColor!;
     final Paint darkPaint = Paint()..color = _config.darkColor!;
 
@@ -390,17 +380,7 @@ class _FlutterLogoPainter extends BoxPainter {
         const Color(0xBF6E6E6E),
         const Color(0xBF616161),
       ],
-      <double>[
-        0.2690,
-        0.4093,
-        0.4972,
-        0.5708,
-        0.6364,
-        0.6968,
-        0.7533,
-        0.8058,
-        0.8219
-      ],
+      <double>[0.2690, 0.4093, 0.4972, 0.5708, 0.6364, 0.6968, 0.7533, 0.8058, 0.8219],
     );
     final Paint trianglePaint = Paint()
       ..shader = triangleGradient
@@ -420,17 +400,7 @@ class _FlutterLogoPainter extends BoxPainter {
         const Color(0x806E6E6E),
         const Color(0x80616161),
       ],
-      <double>[
-        0.4588,
-        0.5509,
-        0.6087,
-        0.6570,
-        0.7001,
-        0.7397,
-        0.7768,
-        0.8113,
-        0.8219
-      ],
+      <double>[0.4588, 0.5509, 0.6087, 0.6570, 0.7001, 0.7397, 0.7768, 0.8113, 0.8219],
     );
     final Paint rectanglePaint = Paint()
       ..shader = rectangleGradient
@@ -503,11 +473,9 @@ class _FlutterLogoPainter extends BoxPainter {
       // only the mark
       logoSize = const Size(202.0, 202.0);
     }
-    final FittedSizes fittedSize =
-        applyBoxFit(BoxFit.contain, logoSize, canvasSize);
+    final FittedSizes fittedSize = applyBoxFit(BoxFit.contain, logoSize, canvasSize);
     assert(fittedSize.source == logoSize);
-    final Rect rect =
-        Alignment.center.inscribe(fittedSize.destination, offset & canvasSize);
+    final Rect rect = Alignment.center.inscribe(fittedSize.destination, offset & canvasSize);
     final double centerSquareHeight = canvasSize.shortestSide;
     final Rect centerSquare = Rect.fromLTWH(
       offset.dx + (canvasSize.width - centerSquareHeight) / 2.0,
@@ -519,8 +487,7 @@ class _FlutterLogoPainter extends BoxPainter {
     Rect logoTargetSquare;
     if (_config._position > 0.0) {
       // horizontal style
-      logoTargetSquare =
-          Rect.fromLTWH(rect.left, rect.top, rect.height, rect.height);
+      logoTargetSquare = Rect.fromLTWH(rect.left, rect.top, rect.height, rect.height);
     } else if (_config._position < 0.0) {
       // stacked style
       final double logoHeight = rect.height * 191.0 / 306.0;
@@ -534,8 +501,7 @@ class _FlutterLogoPainter extends BoxPainter {
       // only the mark
       logoTargetSquare = centerSquare;
     }
-    final Rect logoSquare =
-        Rect.lerp(centerSquare, logoTargetSquare, _config._position.abs())!;
+    final Rect logoSquare = Rect.lerp(centerSquare, logoTargetSquare, _config._position.abs())!;
 
     if (_config._opacity < 1.0) {
       canvas.saveLayer(
@@ -550,23 +516,19 @@ class _FlutterLogoPainter extends BoxPainter {
     if (_config._position != 0.0) {
       if (_config._position > 0.0) {
         // horizontal style
-        final double fontSize =
-            2.0 / 3.0 * logoSquare.height * (1 - (10.4 * 2.0) / 202.0);
+        final double fontSize = 2.0 / 3.0 * logoSquare.height * (1 - (10.4 * 2.0) / 202.0);
         final double scale = fontSize / 100.0;
-        final double
-            finalLeftTextPosition = // position of text in rest position
+        final double finalLeftTextPosition = // position of text in rest position
             (256.4 / 820.0) *
                     rect
                         .width - // 256.4 is the distance from the left edge to the left of the F when the whole logo is 820.0 wide
                 (32.0 / 350.0) *
                     fontSize; // 32 is the distance from the text bounding box edge to the left edge of the F when the font size is 350
-        final double
-            initialLeftTextPosition = // position of text when just starting the animation
+        final double initialLeftTextPosition = // position of text when just starting the animation
             rect.width / 2.0 - _textBoundingRect!.width * scale;
         final Offset textOffset = Offset(
           rect.left +
-              ui.lerpDouble(initialLeftTextPosition, finalLeftTextPosition,
-                  _config._position)!,
+              ui.lerpDouble(initialLeftTextPosition, finalLeftTextPosition, _config._position)!,
           rect.top + (rect.height - _textBoundingRect!.height * scale) / 2.0,
         );
         canvas.save();
@@ -585,8 +547,7 @@ class _FlutterLogoPainter extends BoxPainter {
         canvas.restore();
       } else if (_config._position < 0.0) {
         // stacked style
-        final double fontSize =
-            0.35 * logoTargetSquare.height * (1 - (10.4 * 2.0) / 202.0);
+        final double fontSize = 0.35 * logoTargetSquare.height * (1 - (10.4 * 2.0) / 202.0);
         final double scale = fontSize / 100.0;
         if (_config._position > -1.0) {
           // This limits what the drawRect call below is going to blend with.
