@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart' hide FlutterLogo;
-import 'package:shared_preferences/shared_preferences.dart';
-import 'package:flutter_ume/util/flutter_logo.dart' as flutterLogo;
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter_ume/core/plugin_manager.dart';
 import 'package:flutter_ume/core/ui/root_widget.dart';
+import 'package:flutter_ume/util/flutter_logo.dart' as flutterLogo;
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../../utils/mock_classes.dart';
 
@@ -30,8 +30,7 @@ void main() {
     });
 
     testWidgets('RootWidget pump widget', (tester) async {
-      PluginManager.instance
-          .registerAll([MockPluggable(), MockPluggableWithStream()]);
+      PluginManager.instance.registerAll([MockPluggable(), MockPluggableWithStream()]);
       final umeRoot = UMEWidget(
           child: MaterialApp(
               home: Scaffold(
@@ -49,7 +48,7 @@ void main() {
       final plugin0 = 'MockPluggable';
       final plugin1 = 'MockPluggableWithStream';
       MethodChannel channel = const MethodChannel('bd_shared_preferences');
-      channel.setMockMethodCallHandler((MethodCall methodCall) async {
+      channel.setMethodCallHandler((MethodCall methodCall) async {
         if (methodCall.method == 'commit') {
           return Map<String, dynamic>.from(methodCall.arguments);
         } else if (methodCall.method == 'getAll') {
@@ -63,11 +62,8 @@ void main() {
         }
       });
 
-      PluginManager.instance.registerAll([
-        MockPluggable(),
-        MockPluggableWithStream(),
-        MockPluggableWithNestedWidget()
-      ]);
+      PluginManager.instance.registerAll(
+          [MockPluggable(), MockPluggableWithStream(), MockPluggableWithNestedWidget()]);
       final umeRoot = UMEWidget(
           child: MaterialApp(
               home: Scaffold(
@@ -78,12 +74,11 @@ void main() {
       await tester.pump(Duration(seconds: 1));
       await tester.pumpAndSettle();
 
-      channel.setMockMethodCallHandler(null);
+      channel.setMethodCallHandler(null);
     });
 
     testWidgets('RootWidget flutter logo drag', (tester) async {
-      PluginManager.instance
-          .registerAll([MockPluggable(), MockPluggableWithStream()]);
+      PluginManager.instance.registerAll([MockPluggable(), MockPluggableWithStream()]);
       final umeRoot = UMEWidget(
           child: MaterialApp(
               home: Scaffold(
@@ -105,8 +100,7 @@ void main() {
     });
 
     testWidgets('RootWidget flutter logo drag', (tester) async {
-      PluginManager.instance
-          .registerAll([MockPluggable(), MockPluggableWithStream()]);
+      PluginManager.instance.registerAll([MockPluggable(), MockPluggableWithStream()]);
       final umeRoot = UMEWidget(
           child: MaterialApp(
               home: Scaffold(
@@ -117,16 +111,14 @@ void main() {
       await tester.pump(Duration(seconds: 1));
       await tester.pumpAndSettle();
 
-      final Offset flutterLogoPosition =
-          tester.getCenter(find.byType(flutterLogo.FlutterLogo));
+      final Offset flutterLogoPosition = tester.getCenter(find.byType(flutterLogo.FlutterLogo));
       await tester.dragFrom(flutterLogoPosition, Offset(-100, -100));
       await tester.pump();
       await tester.pumpAndSettle();
     });
 
     testWidgets('RootWidget flutter logo tap', (tester) async {
-      PluginManager.instance
-          .registerAll([MockPluggable(), MockPluggableWithStream()]);
+      PluginManager.instance.registerAll([MockPluggable(), MockPluggableWithStream()]);
       final umeRoot = UMEWidget(
           child: MaterialApp(
               home: Scaffold(
@@ -137,8 +129,7 @@ void main() {
       await tester.pump(Duration(seconds: 1));
       await tester.pumpAndSettle();
 
-      final Offset flutterLogoPosition =
-          tester.getCenter(find.byType(flutterLogo.FlutterLogo));
+      final Offset flutterLogoPosition = tester.getCenter(find.byType(flutterLogo.FlutterLogo));
       await tester.tapAt(flutterLogoPosition);
       await tester.pump();
       await tester.pumpAndSettle();
@@ -151,8 +142,7 @@ void main() {
     testWidgets('RootWidget actions', (tester) async {
       WidgetsFlutterBinding.ensureInitialized();
 
-      PluginManager.instance
-          .registerAll([MockPluggable(), MockPluggableWithStream()]);
+      PluginManager.instance.registerAll([MockPluggable(), MockPluggableWithStream()]);
 
       final umeRoot = UMEWidget(
           child: MaterialApp(
@@ -163,13 +153,11 @@ void main() {
       await tester.pumpWidget(umeRoot);
       await tester.pump(Duration(seconds: 1));
 
-      final Offset flutterLogoPosition =
-          tester.getCenter(find.byType(flutterLogo.FlutterLogo));
+      final Offset flutterLogoPosition = tester.getCenter(find.byType(flutterLogo.FlutterLogo));
       await tester.tapAt(flutterLogoPosition);
       await tester.pump(Duration(seconds: 1));
 
-      final Offset maximalBtnPosition =
-          tester.getCenter(find.byWidgetPredicate((widget) {
+      final Offset maximalBtnPosition = tester.getCenter(find.byWidgetPredicate((widget) {
         return widget is InkWell &&
             (widget.child is CircleAvatar) &&
             (widget.child as CircleAvatar).backgroundColor == Color(0xff53c22b);
@@ -183,8 +171,7 @@ void main() {
       await tester.tapAt(flutterLogoPosition);
       await tester.pump(Duration(seconds: 1));
 
-      final Offset minimalBtnPosition =
-          tester.getCenter(find.byWidgetPredicate((widget) {
+      final Offset minimalBtnPosition = tester.getCenter(find.byWidgetPredicate((widget) {
         return widget is InkWell &&
             (widget.child is CircleAvatar) &&
             (widget.child as CircleAvatar).backgroundColor == Color(0xffe6c029);
@@ -192,8 +179,7 @@ void main() {
       await tester.tapAt(minimalBtnPosition);
       await tester.pump(Duration(seconds: 1));
 
-      final Offset closeBtnPosition =
-          tester.getCenter(find.byWidgetPredicate((widget) {
+      final Offset closeBtnPosition = tester.getCenter(find.byWidgetPredicate((widget) {
         return widget is InkWell &&
             (widget.child is CircleAvatar) &&
             (widget.child as CircleAvatar).backgroundColor == Color(0xffff5a52);
@@ -214,8 +200,7 @@ void main() {
       await tester.pump(Duration(seconds: 1));
       await tester.pumpAndSettle();
 
-      final Offset flutterLogoPosition =
-          tester.getCenter(find.byType(flutterLogo.FlutterLogo));
+      final Offset flutterLogoPosition = tester.getCenter(find.byType(flutterLogo.FlutterLogo));
       await tester.tapAt(flutterLogoPosition);
       await tester.pump();
       await tester.pumpAndSettle();
